chore: refactoring package structure
