+++ src/common/extension/mod.rs
@@ -1,0 +1,7 @@
+pub mod trng;
+
+#[derive(Debug, Clone)]
+pub struct Extension {
+    pub filename: String,
+    pub symbol: String,
+}
