+++ src/common/utils.rs
@@ -1,0 +1,15 @@
+pub mod json {
+    use serde_json::Value;
+
+    pub fn merge(a: &mut Value, b: Value) {
+        match (a, b) {
+            (a @ &mut Value::Object(_), Value::Object(b)) => {
+                let a = a.as_object_mut().unwrap();
+                for (k, v) in b {
+                    merge(a.entry(k).or_insert(Value::Null), v);
+                }
+            }
+            (a, b) => *a = b,
+        }
+    }
+}
