--- src/nodex/extension/mod.rs
@@ -1,1 +1,0 @@
-pub mod trng;
