--- src/nodex/mod.rs
@@ -1,7 +1,0 @@
-pub mod cipher;
-pub mod extension;
-pub mod keyring;
-pub mod runtime;
-pub mod schema;
-pub mod sidetree;
-pub mod utils;
