--- src/nodex/schema/mod.rs
@@ -1,1 +1,0 @@
-pub mod general;
