+++ src/verifiable_credentials/mod.rs
@@ -1,0 +1,3 @@
+pub mod credential_signer;
+pub mod did_vc;
+pub mod types;
